pyproject.toml
[tool.poetry]
name = "estadias_base"
version = "0.1.0"
description = "Django base for estadiasyviajes.com"
authors = ["Carlos Valdemar Soria <carlosvsoria1982@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1"
python-dotenv = "^0.20.0"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^6.3.0"}
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
